Fix reload, move router push addition to individual file
